Reorganize script structure
Importing nifxml.py becomes a little more annoying, but having a subfolder for tools and scripts is necessary for any future expansion (other languages, etc).

Everything that should be available from the nifxml module is imported into __init__.py using an export decorator and __all__ in nifxml.py.

nifdoc.py is a script and not part of package/module so I do not put it inside of `tools/nifdoc`.   I have also left it in the root for now.
