+++ tools/nifxml/__init__.py
@@ -1,0 +1,5 @@
+"""
+All the publicly available imports from __all__
+"""
+
+from .nifxml import * # pylint: disable=W0401
